avoid circular references
