+++ cpp/include/ray/api/ray_runtime.h
@@ -43,5 +43,10 @@ class RayRuntime {
   virtual ObjectID CallActor(const RemoteFunctionPtrHolder &fptr, const ActorID &actor,
                              std::vector<std::unique_ptr<::ray::TaskArg>> &args) = 0;
 };
+
+static std::shared_ptr<RayRuntime> runtime_ = nullptr;
+inline static std::shared_ptr<msgpack::sbuffer> Get(const ObjectID &id) {
+  return runtime_->Get(id);
+}
 }  // namespace api
 }  // namespace ray
